Move errors over to query module
