+++ aquadoggo/src/db/query/mod.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+pub mod errors;
 mod field;
 mod filter;
 mod order;
